+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/Vozhlak/go-sprint-seven-testify
+
+go 1.23.2
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.10.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
